requirements.txt
flask==1.1.2
flask-sqlalchemy
mysqlclient
flask-script
requests
uwsgi
pypinyin
twilio
